Error Checking Request Sent
This error has been fixed a few times during merges and keeps reverting back.  I am fixing it in all branches now to make sure it does not continue to reoccur in future merges.  This fix checks that the request message was successfully sent before setting the pending flag to wait for a reply.