Add start of sethostname.service
Sethostname.service is a one-shot service that was run when setting up the initial RIAPS configured beaglebone black download image.  So, the images will have a host name, but it will be associated with the beaglebone black hardware used to configure the download image.  This image will be placed on different hardware, so the hostname needs to match the current hardware that the image is used.  So, the service needs to be run on each reboot of the system.